feat(sl-1myj): extract shared VizModel types into @satsuma/viz-model package
Create tooling/satsuma-viz-model/ as the canonical home for the VizModel
protocol types shared between the LSP server (producer) and the viz web
component (consumer). Use the LSP version as the superset — it includes
atRefs on TransformInfo, spreads on FragmentCard, and CONSTRAINT_TAGS.

Both satsuma-viz and vscode-satsuma/server now depend on @satsuma/viz-model
and re-export from it; local duplicate definitions are removed. The "keep in
sync" comment is gone — the shared package enforces the contract at build time.

VizModel fixture validation tests move to satsuma-viz-model/test/model.test.js;
the duplicate coverage in satsuma-viz/test/model.test.js is removed.
