+++ tooling/satsuma-viz-model/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@satsuma/viz-model",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared VizModel protocol types — the contract between the Satsuma LSP server and the viz web component",
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "tsc",
+    "prepare": "tsc",
+    "test": "node --test test/**/*.test.js",
+    "pretest": "tsc"
+  },
+  "devDependencies": {
+    "@types/node": "^25.5.0",
+    "typescript": "^6.0.2"
+  }
+}
