+++ tooling/satsuma-viz/package.json
@@ -13,6 +13,7 @@
     "pretest": "tsc --noEmit"
   },
   "dependencies": {
+    "@satsuma/viz-model": "file:../satsuma-viz-model",
     "elkjs": "^0.11.1",
     "lit": "^3.3.0"
   },
