+++ tooling/satsuma-viz/test/model.test.js
@@ -88,123 +88,6 @@ describe("@satsuma/viz bundle", () => {
   });
 });
 
-describe("VizModel fixture validation", () => {
-  it("accepts a minimal VizModel shape", () => {
-    /** @type {import("../src/model.js").VizModel} */
-    const model = {
-      uri: "file:///test.stm",
-      fileNotes: [],
-      namespaces: [
-        {
-          name: null,
-          schemas: [
-            {
-              id: "customers",
-              qualifiedId: "customers",
-              kind: "schema",
-              label: "CRM customers",
-              fields: [
-                {
-                  name: "id",
-                  type: "UUID",
-                  constraints: ["pk", "required"],
-                  notes: [],
-                  comments: [],
-                  children: [],
-                  location: { uri: "file:///test.stm", line: 3, character: 2 },
-                },
-                {
-                  name: "email",
-                  type: "STRING",
-                  constraints: ["pii"],
-                  notes: [],
-                  comments: [
-                    {
-                      kind: "warning",
-                      text: "Must be hashed before export",
-                      location: { uri: "file:///test.stm", line: 4, character: 2 },
-                    },
-                  ],
-                  children: [],
-                  location: { uri: "file:///test.stm", line: 5, character: 2 },
-                },
-              ],
-              notes: [],
-              comments: [],
-              metadata: [],
-              location: { uri: "file:///test.stm", line: 1, character: 0 },
-              hasExternalLineage: false,
-            },
-          ],
-          mappings: [],
-          metrics: [],
-          fragments: [],
-        },
-      ],
-    };
-
-    assert.equal(model.namespaces.length, 1);
-    assert.equal(model.namespaces[0].schemas[0].fields.length, 2);
-    assert.equal(model.namespaces[0].schemas[0].fields[1].constraints[0], "pii");
-    assert.equal(model.namespaces[0].schemas[0].fields[1].comments[0].kind, "warning");
-  });
-
-  it("accepts a metric VizModel shape", () => {
-    /** @type {import("../src/model.js").MetricCard} */
-    const metric = {
-      id: "monthly_revenue",
-      qualifiedId: "monthly_revenue",
-      label: "MRR",
-      source: ["orders"],
-      grain: "monthly",
-      slices: ["segment", "region"],
-      filter: null,
-      fields: [
-        {
-          name: "value",
-          type: "DECIMAL(14,2)",
-          measure: "additive",
-          notes: [],
-          location: { uri: "file:///test.stm", line: 20, character: 2 },
-        },
-        {
-          name: "avg_order",
-          type: "DECIMAL(10,2)",
-          measure: "non_additive",
-          notes: [],
-          location: { uri: "file:///test.stm", line: 21, character: 2 },
-        },
-      ],
-      notes: [],
-      comments: [],
-      location: { uri: "file:///test.stm", line: 18, character: 0 },
-    };
-
-    assert.equal(metric.fields.length, 2);
-    assert.equal(metric.fields[0].measure, "additive");
-    assert.equal(metric.slices.length, 2);
-  });
-
-  it("accepts a fragment VizModel shape", () => {
-    /** @type {import("../src/model.js").FragmentCard} */
-    const fragment = {
-      id: "audit_fields",
-      fields: [
-        {
-          name: "created_at",
-          type: "TIMESTAMP",
-          constraints: ["required"],
-          notes: [],
-          comments: [],
-          children: [],
-          location: { uri: "file:///test.stm", line: 30, character: 2 },
-        },
-      ],
-      notes: [],
-      location: { uri: "file:///test.stm", line: 28, character: 0 },
-    };
-
-    assert.equal(fragment.fields.length, 1);
-    assert.equal(fragment.id, "audit_fields");
-  });
-});
+// VizModel fixture validation tests live in @satsuma/viz-model/test/model.test.js.
+// Duplicating them here would re-test the same invariants after a re-export shim,
+// which adds no value — see satsuma-viz-model for the canonical contract tests.
