+++ tooling/vscode-satsuma/server/package.json
@@ -14,13 +14,14 @@
   },
   "dependencies": {
     "@satsuma/core": "file:../../satsuma-core",
+    "@satsuma/viz-model": "file:../../satsuma-viz-model",
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12",
     "web-tree-sitter": "^0.26.7"
   },
   "devDependencies": {
     "@types/node": "^25.5.0",
-    "typescript": "^6.0.2",
-    "esbuild": "^0.27.4"
+    "esbuild": "^0.27.4",
+    "typescript": "^6.0.2"
   }
 }
